+++ designer/submodules/packages/components/src/components/utils/index.ts
@@ -1,3 +1,2 @@
 export { Truncate } from "./truncate";
 export { TruncateWrapper } from "./truncateWrapper";
-export { useCellArrowKeys } from "./useCellArrowKeys";